agrego de la funcion eliminar al area de fondo

+++ Sistema/Fondos/Gestor_tbl_Donante.php
@@ -103,7 +103,7 @@ if ($datos=$sql->fetch_object()) {
 }
 $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol and ID_Objeto=6");
 if ($datos=$sql->fetch_object()) { 
-        $output['data'] .= '<td><a class="boton-eliminar" href="Delete_Donan.php?ID_Donante=' . $row['ID_Donante'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
+        $output['data'] .= '<td><a onclick="return confirmar()" class="boton-eliminar" href="Delete_Donan.php?ID_Donante=' . $row['ID_Donante'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
 }
 $output['data'] .= '</tr>';
     }
