+++ Sistema/Fondos/Gestor_tbl_Tipo_de_Fondos.php
@@ -100,7 +100,7 @@ if ($datos=$sql->fetch_object()) {
 }
 $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol and ID_Objeto=13");
 if ($datos=$sql->fetch_object()) { 
-        $output['data'] .= '<td><a class="boton-eliminar" href="Delete_Tipo_Fondo.php?ID_tipo_fondo=' . $row['ID_tipo_fondo'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
+        $output['data'] .= '<td><a onclick="return confirmar()" class="boton-eliminar" href="Delete_Tipo_Fondo.php?ID_tipo_fondo=' . $row['ID_tipo_fondo'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
 }
         $output['data'] .= '</tr>';
     }
